requirements.txt
plotly-express
pandas
datacommons_pandas
kaleido